[percentile] move gk_array into the percentile package
