+++ pkg/aggregator/sketch_series.go
@@ -1,5 +1,9 @@
 package aggregator
 
+import (
+	"github.com/DataDog/datadog-agent/pkg/percentile"
+)
+
 // Sketch represents a quantile sketch at a specific time
 type Sketch struct {
 	Timestamp int64   `json:"timestamp"`
@@ -19,12 +23,12 @@ type SketchSerie struct {
 // QSketch is a wrapper around GKArray to make it easier if we want to try a
 // different sketch algorithm
 type QSketch struct {
-	GKArray
+	percentile.GKArray
 }
 
 // NewQSketch creates a new QSketch
 func NewQSketch() QSketch {
-	return QSketch{NewGKArray()}
+	return QSketch{percentile.NewGKArray()}
 }
 
 // NoSketchError is the error returned when not enough samples have been
